Azure_Nba_App/requirements.txt
langchain
openai
langchain_openai
azure-functions
